+++ examples/java8/pom.xml
@@ -145,6 +145,11 @@
 
     <dependency>
       <groupId>org.apache.beam</groupId>
+      <artifactId>beam-sdks-java-io-google-cloud-platform</artifactId>
+    </dependency>
+
+    <dependency>
+      <groupId>org.apache.beam</groupId>
       <artifactId>beam-examples-java</artifactId>
       <version>${project.version}</version>
     </dependency>
